move announcements and holidays to a separate app
