+++ corevolt/settings.py
@@ -58,6 +58,7 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',
     'corsheaders',
     'corevolthrm',
+    'home',
 ]
 
 MIDDLEWARE = [
