+++ corevolt/urls.py
@@ -20,4 +20,5 @@ from django.urls import path,include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('corevolthrm.urls')),
+    path('', include('home.urls')),
 ]
