+++ corevolthrm/admin.py
@@ -1,9 +1,4 @@
 from django.contrib import admin
 from .models import CustomUser
 from django.contrib.auth.admin import UserAdmin
-from corevolthrm.models import Announcement, Holiday
 admin.site.register(CustomUser)
-
-# Register your models here.
-admin.site.register(Announcement)
-admin.site.register(Holiday)
