+++ corevolthrm/models.py
@@ -28,18 +28,3 @@ class CustomUser(AbstractUser):
     
     def __str__(self):
         return self.email
-    
-class Announcement(models.Model):
-    created = models.DateTimeField(auto_now_add=True)
-    title = models.CharField(max_length=100, blank=False)
-    content = models.TextField()
-
-    class Meta:
-        ordering = ['-created']
-
-class Holiday(models.Model):
-    name = models.CharField(max_length = 30, blank=False)
-    date = models.DateField()
-
-    class Meta:
-        ordering = ['date']
