+++ corevolthrm/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 from django.contrib.auth import get_user_model
 from django.contrib.auth.password_validation import validate_password
-from corevolthrm.models import Announcement, Holiday
+
 
 User = get_user_model()
 
@@ -27,18 +27,3 @@ class UserRegistrationSerializer(serializers.ModelSerializer):
         validated_data.pop('password2')
         user = User.objects.create_user(**validated_data)
         return user
-
-class AnnouncementSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Announcement
-        fields = '__all__'
-
-class HolidaySerializer(serializers.ModelSerializer):
-    day_of_the_week = serializers.SerializerMethodField()
-
-    def get_day_of_the_week(self, obj):
-        return obj.date.strftime("%A")
-
-    class Meta:
-        model = Holiday
-        fields = ['name', 'date', 'day_of_the_week']
