+++ corevolthrm/views.py
@@ -15,9 +15,6 @@ from rest_framework.decorators import api_view,permission_classes
 from rest_framework_simplejwt.tokens import RefreshToken
 from rest_framework.permissions import IsAuthenticated
 
-from corevolthrm.models import Announcement, Holiday
-from corevolthrm.serializers import AnnouncementSerializer, HolidaySerializer
-
 # Create your views here.
 def profile_list(request):
     if request.method == 'GET':
@@ -144,13 +141,3 @@ def logoutUser(request):
     response.delete_cookie('access_token')
     response.delete_cookie('refresh_token')
     return response
-
-class AnnouncementList(generics.ListCreateAPIView):
-    queryset = Announcement.objects.all()
-    serializer_class = AnnouncementSerializer
-    permission_classes = [IsAuthenticated]
-
-class HolidayList(generics.ListAPIView):
-    queryset = Holiday.objects.all()
-    serializer_class = HolidaySerializer
-    # permission_classes = [IsAuthenticated]
