+++ home/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from home.models import Announcement, Holiday
+
+admin.site.register(Announcement)
+admin.site.register(Holiday)
